Create InactiveTabsTest.swift
