+++ firefox-ios/firefox-ios-tests/Tests/PerformanceTestPlan.xctestplan
@@ -42,6 +42,7 @@
         "HistoryTests",
         "HomeButtonTests",
         "HomePageSettingsUITests",
+        "InactiveTabsTest",
         "IntegrationTests",
         "IpadOnlyTestCase",
         "IphoneOnlyTestCase",
