+++ firefox-ios/firefox-ios-tests/Tests/Smoketest2.xctestplan
@@ -58,6 +58,7 @@
         "HistoryTests",
         "HomeButtonTests",
         "HomePageSettingsUITests",
+        "InactiveTabsTest",
         "IntegrationTests",
         "IpadOnlyTestCase",
         "IphoneOnlyTestCase",
