+++ firefox-ios/firefox-ios-tests/Tests/Smoketest4.xctestplan
@@ -85,6 +85,7 @@
         "HistoryTests\/testTabHistory()",
         "HomeButtonTests",
         "HomePageSettingsUITests",
+        "InactiveTabsTest",
         "IntegrationTests",
         "IpadOnlyTestCase",
         "IphoneOnlyTestCase",
