+++ firefox-ios/firefox-ios-tests/Tests/SyncIntegrationTestPlan.xctestplan
@@ -63,6 +63,7 @@
         "HistoryTests",
         "HomeButtonTests",
         "HomePageSettingsUITests",
+        "InactiveTabsTest",
         "IpadOnlyTestCase",
         "IphoneOnlyTestCase",
         "JumpBackInTests",
